카트 등록 및 즉시구매 api 작성 완료.
:)

+++ src/main/java/bashpound/marketplace/infra/repository/CartMapper.java
@@ -10,7 +10,7 @@ public interface CartMapper {
 
 	public List<Cart> getCart(@Param("username") String username);
 
-	public int insert(@Param("prodId") Long pid, @Param("purchaseId") int purchaseId,
+	public int insert(@Param("prodId") Long pid, @Param("purchaseId") Long purchaseId,
 			@Param("numberOfItems") int numberOfItems);
 
 }
